pamparam.el: Split cards/ directory into buckets
Each bucket is the first two letters of the card name.

The reasoning: it takes a huge amount of time to open a dired buffer
with thousands of files in it.

Not that it's super-useful to open a card's directory; but sometimes
it's interesting to look at the last modification date, or file size
etc.