pamparam.el (pam-commit): Count more files
When a card that's scheduled for today gets updated, it's not seen as
"modified" in Git's history. Instead, it's deleted and a new one with
similar content is added and scheduled in place of the old one.

Therefore, count new card files for purposes of reporting the amount of
cards learned. One disadvantage is that actually new cards would be
counted as well. A simple solution to that is call `pam-commit' twice.

A different approach to counting done cards would be to compute the
difference between DONE items in the old and new version of today's
schedule file.